requirements.txt
discord.py==1.2.3
youtube-dl==2019.8.2
PyNaCl==1.3.0